refactor: Remove statusBarNotifications from grid item parameters (folder-bg-pad)
This commit removes the `statusBarNotifications` parameter from various grid item composables and their callers.
The logic for displaying notification badges has been moved to `InteractiveGridItem.kt` and `InteractiveFolderGridItem.kt`
where it is handled directly within the composable, simplifying the parameter passing.

The following files were affected:
- `feature/home/src/main/kotlin/com/eblan/launcher/feature/home/screen/editpage/EditPageScreen.kt`
- `feature/home/src/main/kotlin/com/eblan/launcher/feature/home/screen/editpage/GridItem.kt`
- `feature/home/src/main/kotlin/com/eblan/launcher/feature/home/screen/folder/InteractiveFolderGridItem.kt`
- `feature/home/src/main/kotlin/com/eblan/launcher/feature/home/screen/pager/InteractiveGridItem.kt`
- `feature/home/src/main/kotlin/com/eblan/launcher/feature/home/HomeScreen.kt`
- `feature/home/src/main/kotlin/com/eblan/launcher/feature/home/screen/pager/PagerScreen.kt`
