+++ feature/home/src/main/kotlin/com/eblan/launcher/feature/home/screen/pager/PagerScreen.kt
@@ -19,14 +19,17 @@ package com.eblan.launcher.feature.home.screen.pager
 
 import android.content.BroadcastReceiver
 import android.content.ClipDescription
+import android.content.ComponentName
 import android.content.Context
 import android.content.Intent
 import android.content.IntentFilter
+import android.content.ServiceConnection
 import android.content.pm.ActivityInfo
 import android.content.pm.LauncherApps
 import android.content.pm.ShortcutInfo
 import android.os.Build
 import android.os.Handler
+import android.os.IBinder
 import android.os.Looper
 import android.os.UserHandle
 import androidx.activity.ComponentActivity
@@ -91,6 +94,7 @@ import androidx.lifecycle.Lifecycle
 import androidx.lifecycle.LifecycleEventObserver
 import androidx.lifecycle.compose.LocalLifecycleOwner
 import androidx.lifecycle.lifecycleScope
+import androidx.lifecycle.repeatOnLifecycle
 import com.eblan.launcher.domain.model.AppDrawerSettings
 import com.eblan.launcher.domain.model.Associate
 import com.eblan.launcher.domain.model.EblanAppWidgetProviderInfo
@@ -131,6 +135,7 @@ import com.eblan.launcher.feature.home.util.SCALE
 import com.eblan.launcher.feature.home.util.calculatePage
 import com.eblan.launcher.feature.home.util.getTextColor
 import com.eblan.launcher.framework.usermanager.AndroidUserManagerWrapper
+import com.eblan.launcher.service.EblanNotificationListenerService
 import com.eblan.launcher.ui.local.LocalAppWidgetHost
 import com.eblan.launcher.ui.local.LocalAppWidgetManager
 import com.eblan.launcher.ui.local.LocalFileManager
@@ -172,7 +177,6 @@ internal fun PagerScreen(
     gridItemSource: GridItemSource?,
     isVisibleOverlay: Boolean,
     previewFolderGridItems: Map<String, PreviewFolder>,
-    statusBarNotifications: Map<String, Int>,
     iconPackInfoFilePaths: Map<String, String?>,
     onDeleteGridItem: (GridItem) -> Unit,
     onResetGridAfterDeleteGridItem: (GridItem) -> Unit,
@@ -478,6 +482,8 @@ internal fun PagerScreen(
         pagerScreenState.showWidgetScreen ||
         pagerScreenState.showShortcutConfigScreen ||
         pagerScreenState.eblanApplicationInfoGroup != null
+
+    val statusBarNotifications by rememberStatusBarNotifications()
 
     LaunchedEffect(
         key1 = pinGridItem,
@@ -1455,6 +1461,52 @@ private fun rememberManagedProfileResult(): State<ManagedProfileResult?> {
 
         awaitDispose {
             context.unregisterReceiver(receiver)
+        }
+    }
+}
+
+@Composable
+private fun rememberStatusBarNotifications(): State<Map<String, Int>> {
+    val context = LocalContext.current
+
+    val lifecycleOwner = LocalLifecycleOwner.current
+
+    return produceState(
+        initialValue = emptyMap(),
+        key1 = context,
+        key2 = lifecycleOwner,
+    ) {
+        val serviceConnection = object : ServiceConnection {
+            override fun onServiceConnected(
+                name: ComponentName,
+                service: IBinder,
+            ) {
+                val listener =
+                    (service as EblanNotificationListenerService.LocalBinder).getService()
+
+                launch {
+                    lifecycleOwner.repeatOnLifecycle(Lifecycle.State.STARTED) {
+                        listener.statusBarNotifications.collect {
+                            value = it
+                        }
+                    }
+                }
+            }
+
+            override fun onServiceDisconnected(name: ComponentName) = Unit
+        }
+
+        context.bindService(
+            Intent(
+                context,
+                EblanNotificationListenerService::class.java,
+            ),
+            serviceConnection,
+            Context.BIND_AUTO_CREATE,
+        )
+
+        awaitDispose {
+            context.unbindService(serviceConnection)
         }
     }
 }
